Move service operations to the the service packages (#7950)
* Move compute_operation.go.erb to the compute package

* Move appengine_operation.go to the appengine package

* Move cloudfunctions_operation.go to the cloudfunctions package

* Move composer_operation.go.erb to the composer package

* Move container_attached_operation.go.erb to the containerattached package

* Move container_operation.go.erb to the container package

* Move dataproc_cluster_operation.go to the dataproc package

* Move dataproc_job_operation.go to the dataproc package

* Move datastream_operation.go to the datastream package

* Fix import path

* Add compute_operation.go.erb in the compute package to the converter

* Fix common~copy

* Move dialogflow_cx_operation.go to the service package

* Move gkeonprem_operation.go to the service package

* Move runadminv3_operation.go to the service package

* Move service_networking_operation.go to the service package

* Move serviceman_operation.go to the service package

* Move sqladmin_operation.go to the service package

* Move tags_location_operation.go to the service package

* Move vertex_ai_operation.go to the service package

* Move deployment_manager_operation.go to the service package

* Add sql/sqladmin_operation.go to the converter copy list

* Test 7982

* Skip build steps

* Remove image test
